Create README.md
My final project for ENME351 warranted me to use Sensing, Logic, and Actuation using Arduino to pass information over to Processing. I recreated the game Space Invaders with a few twists to meet my courses requirements. I included the following to demonstrate my understand of the key principles of the course: 

- A photoresistor to notify the user of their environment (light vs. dark)
- A button for the user to control light/dark mode in the Processing module
- A potentiometer to drag the turret left and right across the screen
- A second button to initiate a laser on the Processing module, and finally 
- A buzzer to play the intro to 7 Nation Army by The White Stripes once the game has concluded. 

The Processing module also displays a GIF (game over screen) to also indicate the end of the game. It was one of the features that made my project unique. Thank you for reading! Hope you enjoy the demonstration.